Editando Funções
Redefinição de atributos

+++ ClimaAvi/Models/Barometro.cs
@@ -4,17 +4,17 @@ namespace ClimaAvi.Models
 {
     public class Barometro
     {
-        public double Altitude { get; set; }
-        public double Temperatura { get; set; }
-        public double PressaoAtmosferica { get; set; }
-        public double UmidadeAr { get; set; }
+        public float Altitude { get; set; }
+        public float Temperatura { get; set; }
+        public float PressaoAtmosferica { get; set; }
+        public float UmidadeAr { get; set; }
         public DateTime Leitura { get; set; }
 
         public Barometro()
         {
         }
 
-        public Barometro(double altitude, double temperatura, double pressaoAtmosferica, double umidadeAr, DateTime leitura)
+        public Barometro(float altitude, float temperatura, float pressaoAtmosferica, float umidadeAr, DateTime leitura)
         {
             Altitude = altitude;
             Temperatura = temperatura;
